Latest tools sync

+++ Tooling/dscbuild/Clear-InstalledDscResource.ps1
@@ -1,10 +1,14 @@
 function Clear-InstalledDscResource {
-    [cmdletbinding()]
+    [cmdletbinding(SupportsShouldProcess=$true)]
     param ()
 
-    $ProgramFilesModules = join-path $env:ProgramFiles 'WindowsPowerShell\Modules'
-    Write-Verbose 'Clearing Program Files from old configuration modules.'
-    dir $script:DscBuildParameters.SourceModuleRoot |
-        Foreach { dir (join-path $ProgramFilesModules $_.Name) -erroraction SilentlyContinue } |
-        Remove-Item -Force -Recurse
+    if ( Test-BuildResource ) {        
+        if ($pscmdlet.shouldprocess("$($script:DscBuildParameters.ProgramFilesModuleDirectory)) based on modules in $($script:DscBuildParameters.SourceResourceDirectory)")) {            
+            Write-Verbose 'Clearing Program Files from old configuration modules.'
+            dir $script:DscBuildParameters.SourceResourceDirectory |
+                Foreach { dir (join-path $script:DscBuildParameters.ProgramFilesModuleDirectory $_.Name) -erroraction SilentlyContinue } |
+                Remove-Item -Force -Recurse
+        }
+    }
 }
+
